feat (layout): Update Navbar
